requirements.txt
# Core ML
tensorflow==2.21.0
numpy>=1.26.4

# Image handling
opencv-python-headless>=4.10.0
Pillow>=10.4.0

# Visualisation / plotting
matplotlib>=3.8.0

# Web app
streamlit>=1.35.0

# PDF generation
fpdf2>=2.7.0
